+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `partial` expector
+
 ## [0.1.0] - 2025-09-19
 
-Initial release.
+- Initial release